+++ test/reports/report_data_test.dart
@@ -15,7 +15,7 @@ import 'package:mutation_test/src/core/commands.dart';
 import 'package:test/test.dart';
 
 import 'create_test_data.dart';
-import '../mock_system_interactions.dart';
+import '../core/mock_system_interactions.dart';
 
 void main() {
   group('With data', () {