+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - Performing a dry run will now generate the specified outputs. All mutations
    defined by the rules will be marked as undetected.
  - Added optional "id" attribute to rules in the xml configuration file
+ - xml configuration file version increased to 1.1
  - Added a column in the generated html report that shows the mutation pattern and id
 
 ## 1.3.4